requirements.txt
PyQt6>=6.4.0
plotly>=5.13.0
python-ptrace>=0.9.8
PyYAML>=6.0
cryptography>=39.0.0
psutil>=5.9.0
pandas>=2.0.0
pytest>=7.3.1
black>=23.3.0
pylint>=2.17.0
